Refactor plugin class
