+++ ckanext/showcase/cli.py
