+++ setup.py
@@ -82,8 +82,6 @@ setup(
         [babel.extractors]
         ckan = ckan.lib.extract:extract_ckan
 
-        [paste.paster_command]
-        showcase=ckanext.showcase.commands.paster:MigrationCommand
     ''',
 
     message_extractors={
